Add default values for session practice details
If the pages that pull practice details from the session are accessed
directly then the page appears to be missing text.

During normal usage, a user should go through a page that sets the
session, so this shouldn't impact them. For cases where we are hitting
that page directly, for example to grab a screenshot, it's nice to have
some default text.