Show waiting time in days, not as % happiness
We've seen "wait time" be interpreted in multiple ways -- is it the
amount of time until the next available appointment, or is it how long
you have to spend sitting in the waiting room?

Showing it like this should remove that ambiguity, and allow people to
provide their own valuation. 8 days for a routine appointment might be
good for some people, and awful for others.