Progressively reveal "armed forces" and "abroad"
Don't show "Have you recently left the armed forces?" upfront - only do
this if they are *not* currently registered with a GP practice.

Don't show "Are you moving to the UK from abroad?" until they say "no"
to having left armed forces.

I'm making the assumption here that it's not possible to be registered
with a GP **and** recently have left the armed forces. Ditto with having
moved back from abroad.