Split out service data from appointments
The information about the service is (hopefully) separate from the
appointment data. We can set the context for the service (ie bring
glasses, your eyes might be blurry afterwards) independently of the
appointment options available.

This is the first step towards pulling this data from the db files and
integrating these service booking flows into the main book an
appointment prototype.