Add diabetes review cards to planner
Rejig how we're building the list. We're still using query parameters to
work out the state, but now add the cards in one at a time.

This avoids having to enumerate all the possible list combinations based
on the query string.

Once we allow not-hardcoded appointments to be booked, the ordering will
probably end up horribly wrong. Classifying that as a problem for
another commit.