First stab at wiring booking into the planner
Add a route to start the eye test booking journey. It sets a session
containing the return URL to be used at the end of the booking.

Change the booking confirmation route to check for the offramp return
URL. If it finds one, redirect to it. If not, render the confirmation
page as usual.

Also clear the session when using the log in/enter your details page, as
it suggests the user hasn't come from a place where we need or want
offramps to work.

Ideally the appointment confirmation card should be open and centred
when the planner is loaded after return, but I'm going to leave that
exercise for a future piece of work.