Remove apply for free prescriptions from planner
In the current world the process starts with a paper form that you have
to get from your GP, and then you give it back to your GP to sign and
send off.

In the ideal world, free prescriptions would Just Work, and the user
wouldn't have to worry about applying.

Since that means there's nothing obvious for us to show when the user
clicks on the button, and we don't think it will prompt any useful
discussions in the lab, we're ditching it completely.