Hide some preference filters for service booking
None of the services we're currently offering obviously present a need
for a gender preference.

For the appointments held in the GP practice offering a choice of
practitioner still makes sense, as you may have an existing relationship
with the people.

For the eye screening appointment held elsewhere and done once a year
offering that choice seems non-useful and likely non-possible.