Update test results target date
The intent of this is a little confusing. We're trying to do a couple of
things at once:

 1. Show the results from your last blood test, and how that compared to
    your target.
 2. Show the target for your next blood test, as discussed with your GP.

Since we don't have time to overhaul the test results card before the
lab, update the card so it's only trying to do the first thing, by
showing the result and target for the same date.