Only present one new appointment per preference
When a user clicks a preference like "I'd like an early morning
appointment", only add one new appointment to their screen, not two.

At some point in the past we changed from presenting two appointments on
the first screen to one but we didn't reduce the number on subsequent pages
from three to two.

This also has the consequence of removing telephone appointments which
are currently a bit confusing.